added new badges
AWS switch to a new system for badges called 'Credly', which conveniently the Linux foundation uses as well. Both the AWS and LF badges have been added. The images are downloaded and have the url in an <a> tag because the embeded html from credly doesn't work properly on Firefox.